Split word-range into class and helpers
